♻️ refactor: 타입 파일 위치 변경 #26
