+++ src/apis/festivals/hotFestival/hotFestivalType.ts
