+++ src/app/(home)/festivals/mostlike/view.tsx
@@ -1,4 +1,4 @@
-import { HostFestivalData } from "@/apis/festivals/hotFestival/types/hotFestival";
+import { HostFestivalData } from "@/apis/festivals/hotFestival/hotFestivalType";
 import { TrendFestivalCard } from "@/components/core/Card";
 
 interface Props {
