+++ src/components/core/Card/TrendFestivalCard/TrendFestivalCard.tsx
@@ -2,7 +2,7 @@ import Image from "next/image";
 import Link, { LinkProps } from "next/link";
 import { FC } from "react";
 
-import { FestivalListModel } from "@/apis/festivals/hotFestival/types/hotFestival";
+import { FestivalListModel } from "@/apis/festivals/hotFestival/hotFestivalType";
 import { DateTag } from "@/components/core/Tag";
 import useDayjs from "@/hooks/useDayjs";
